+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KK426SRXA92Y85G3QY1FNBWG"
+        "01KK7DZPAMGPS257Z5P1DZWM6B"
     )
     r"""The unique identifier of the data source"""