+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KK426SRTCK6EZNH6M9HNRGVA"
+        "01KK7DZPAJBZ6W09EWQCZ8N9AR"
     )
     r"""The unique identifier of the data source"""