+++ packages/orq-rc/RELEASES.md
@@ -6359,3 +6359,13 @@ Based on:
 - [python v4.5.0-rc.38] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.38] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.38 - packages/orq-rc
+
+## 2026-03-08 19:17:28
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.749.0 (2.855.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.39] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.39] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.39 - packages/orq-rc